server/requirements.txt
numpy
requests
flask
pillow
